Domestic_XHS_Project/requirements.txt
# Browser automation (OpenCLI core)
playwright>=1.44.0
playwright-stealth>=1.0.6

# AI
anthropic>=0.31.0

# Configuration & env
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.7.0

# Logging
structlog>=24.1.0
colorama>=0.4.6

# Data processing
pandas>=2.2.0

# Scheduling
croniter>=2.0.0

# HTTP
httpx>=0.27.0

# NLP (lightweight)
jieba>=0.42.1

# Proxy management
proxy-manager>=1.0.0

# Utilities
diskcache>=5.6.0
rich>=13.0.0
